Separate pipeline definition from run functions
