+++ app/Commands/Dev/Highlight.hs
@@ -3,6 +3,7 @@ module Commands.Dev.Highlight where
 import Commands.Base
 import Commands.Dev.Highlight.Options
 import Juvix.Compiler.Concrete.Data.Highlight qualified as Highlight
+import Juvix.Compiler.Pipeline.Run
 
 runCommand :: (Members '[Embed IO, App] r) => HighlightOptions -> Sem r ()
 runCommand HighlightOptions {..} = do
